Multi-Agent/pyproject.toml
[project]
name = "multi-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-generativeai>=0.8.5",
    "ipykernel>=6.29.5",
    "langgraph>=0.5.3",
    "python-dotenv>=1.1.1",
]
